backend/ai/requirements.txt
Flask==2.2.2
python-dotenv
boto3

celery
opencv-python
fer
Pillow
numpy
pandas
tensorflow>=1.7
gunicorn